revisado el server

+++ src/server/UpdateRace.cpp
@@ -9,10 +9,10 @@ UpdateRace::UpdateRace(int32_t client_ID, int32_t key, int32_t key_state) {
     this->key_state = key_state;
 }
 
-void UpdateRace::update_cars(std::unordered_map<uint8_t, Car>& cars) {
+void UpdateRace::update_cars(std::unordered_map<int32_t, Car>& cars) {
     if (this->key_state == KEY_PRESSED) {
         cars.at(this->client_ID).press_key(key);
-    } else {
+    } else /*KEY_UNPRESSED*/ {
         cars.at(this->client_ID).release_key(key);
     }
 }
